+++ dev/index.js
@@ -1,9 +1,11 @@
 import PhyloCanvas, * as phyloComponents from '../src/index';
 import historyPlugin from '../src/plugins/History';
 import ajaxPlugin from '../src/plugins/Ajax';
+import metadataPlugin from '../src/plugins/Metadata';
 
 PhyloCanvas.plugin(historyPlugin);
 PhyloCanvas.plugin(ajaxPlugin);
+PhyloCanvas.plugin(metadataPlugin);
 
 let buttonForm = document.getElementById('buttons');
 let tree = PhyloCanvas.createTree('phylocanvas', {
@@ -39,7 +41,12 @@ tree.on('error', function (event) { throw event.error; });
 
 tree.on('loaded', function () {
   console.log('loaded');
-  // tree.viewMetadataColumns();
+  for (let i = 0; i <= 12; i++) {
+    if (tree.leaves[i]) {
+      tree.leaves[i].data = { col: 1, x: 0, a: 1, c: 0 };
+    }
+  }
+  tree.viewMetadataColumns();
 });
 
 tree.alignLabels = true;
